:art: Tweak install function a bit
